requirements.txt
python-telegram-bot==12.8
telethon
pyrogram
aiohttp[speedups]>=3.7.2
bleach
emoji
future
google_trans_new
nltk
psutil
fontTools
markdown2
Pillow
PyYAML
requests
spamwatch
sqlalchemy
wikipedia
bs4
speedtest-cli
coffeehouse
jikanpy
zalgo-text
pyrate-limiter
psycopg2-binary
hurry.filesize
alphabet_detector
beautifulsoup4
regex
lxml
TgCrypto
cachetools
telegraph
CurrencyConverter
spongemock
geopy
ffmpeg
